pyproject.toml
[project]
name = "crucible-upload-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0",
    "nano-crucible>=2.0.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "qrcode>=8.2",
]
